// web-ui/client/gen/sales/contact.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

export interface Contact {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  /** @maxLength 254 */
  email: string;
  /** @maxLength 20 */
  phone?: string;
  account: number;
  readonly account_name: string;
  /** @maxLength 255 */
  title?: string;
  readonly created_at: string;
}