// web-ui/client/gen/sales/account.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

export interface Account {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  /** @maxLength 255 */
  industry?: string;
  /** @maxLength 100 */
  size?: string;
  /** @maxLength 255 */
  location?: string;
  /** @maxLength 200 */
  website?: string;
  readonly created_at: string;
}